+++ src/Concerto/PanelBundle/Service/DataTableService.php
@@ -297,12 +297,12 @@ class DataTableService extends AExportableSectionService
         }
     }
 
-    public function saveColumn($object_id, $column_name, $new_name, $new_type)
+    public function saveColumn($object_id, $column_name, $new_name, $new_type, $nullable = false)
     {
         $object = $this->get($object_id);
         if ($object != null) {
             $this->update($object);
-            return $this->dbStructureService->saveColumn($object->getName(), $column_name, $new_name, $new_type);
+            return $this->dbStructureService->saveColumn($object->getName(), $column_name, $new_name, $new_type, $nullable);
         } else {
             return array();
         }
@@ -572,14 +572,14 @@ class DataTableService extends AExportableSectionService
             foreach ($old_columns as $old_col) {
                 if ($old_col["name"] == $new_col["name"]) {
                     $found = true;
-                    $db_errors = $this->dbStructureService->saveColumn($new_name, $old_col["name"], $new_col["name"], $new_col["type"]);
+                    $db_errors = $this->dbStructureService->saveColumn($new_name, $old_col["name"], $new_col["name"], $new_col["type"], $new_col["nullable"]);
                     if (count($db_errors) > 0)
                         return array("errors" => $db_errors, "entity" => null, "source" => $obj);
                     break;
                 }
             }
             if (!$found) {
-                $db_errors = $this->dbStructureService->saveColumn($new_name, "0", $new_col["name"], $new_col["type"]);
+                $db_errors = $this->dbStructureService->saveColumn($new_name, "0", $new_col["name"], $new_col["type"], $new_col["nullable"]);
                 if (count($db_errors) > 0)
                     return array("errors" => $db_errors, "entity" => null, "source" => $obj);
             }